lexer: split wasm parts into a separate crate
This should avoid cache invalidation during cargo test.
