+++ bin/lint
@@ -45,7 +45,7 @@ copyright_files=$(grep -vE \
     -e '(^|/)(Cargo|askama)\.toml$' \
     -e '^\.cargo/config$' \
     -e '^\.config/hakari.toml$' \
-    -e '^Cargo\.lock$' \
+    -e '(^|/)Cargo\.lock$' \
     -e '^about\.toml$' \
     -e '^deny\.toml$' \
     -e '^netlify\.toml$' \
