+++ src/sql-lexer-wasm/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "mz-sql-lexer-wasm"
+description = "The lexer for Materialize's SQL dialect, with wasm build targets."
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+mz-sql-lexer = { path = "../sql-lexer", default-features = false }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen = "0.2.63"
+lol_alloc = "0.4.0"
+
+[package.metadata.wasm-pack.profile.release]
+# Optimize for small code size. Verified this is better than 'Os'
+wasm-opt = ['-Oz']
