+++ src/sql-lexer/Cargo.toml
@@ -6,19 +6,12 @@ edition.workspace = true
 rust-version.workspace = true
 publish = false
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
 [dependencies]
 mz-ore = { path = "../ore", default-features = false, features = ["stack"] }
 workspace-hack = { version = "0.0.0", path = "../workspace-hack", optional = true }
 serde = { version = "1.0.152", features = ["derive"] }
 phf = { version = "0.11.1", features = ["uncased"] }
 uncased = "0.9.7"
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen = "0.2.63"
-lol_alloc = "0.4.0"
 
 [dev-dependencies]
 mz-ore = { path = "../ore", default-features = false, features = ["test"] }
@@ -35,7 +28,3 @@ default = ["workspace-hack"]
 
 [package.metadata.cargo-udeps.ignore]
 normal = ["workspace-hack"]
-
-[package.metadata.wasm-pack.profile.release]
-# Optimize for small code size. Verified this is better than 'Os'
-wasm-opt = ['-Oz']
